Restore running in DPI unaware mode
An earlier change caused winspy to run in per-monitor DPI aware mode.
This means that the system never automatically scales up the window
contents when running on a high DPI system.  This makes the tool
unusably tiny on truly high DPI systems (say, >= 250% DPI).

Substantial work is required to make the tool properly support
per-monitor DPI aware.  Logic needs to be added to resize/re-layout
things when the window is moved to a monitor of a different DPI, etc.

For now, I'm putting this back into unaware mode (as it used to be).
When there is no DPI awareness entry in the manifest, the app will
defualt to DPI unaware mode.

The UI of the tool itself will be a bit blurry on high DPI systems
but that is better than being so tiny that you cannot read the text.