Show keyboard focus rects for the bitmap-only buttons
Image-only bitmap buttons did not draw a focus rect.  The code was
explicit about this, presumably as an asthetic choice.

However, I think that this was a bad call.  This makes keyboard based
navigation of the app difficult since there is no visible indication
when you've tabbed onto one of these buttons.

I made these buttons draw the focus rect in all cases.  Mouse only
users won't see any difference.

Also, fixed a focus order issue on the Process tab.  The arrow button
to the right of the module name field should follow it in the focus
order.  This is consistent with how the flow on the General tab works.