Change window-flash to use an overlay window instead of xor
The old flash window code worked by repeatedly xor'ing some
rects on the screen DC.  This worked okay in the old days, but
on the more recent versions of windows (Win10 era) this leaves
garbage on the screen.

I change the code to use a transparent overlay window much
like how the window picker works (which itself used to be built
around xor'ing rects onto the screen DC).  The window just toggles
its visiblity on/off for a second or so to generate the flashing
effect.

Currently, this uses the same red image as the window picker.
Looks okay, though perhaps a bit glaring for larger window.
This can be refined.