Support '0x' prefix on hex inputs
The _tstrtoib16 helper parses hex nubmers.  Updated it to allow for
a leading 0x prefix and leading whitespace.  e.g. all of these are
now treated as equal:

  "1234", "0x1234", "  1234", " 0x1234"

This makes it nicer to copy paste HWND values from a debugger and
paste them into the 'Handle' field on the General tab.