Improve treeview icons, mainly for hidden windows
Some changes to the imagelist strips that are used for the
treeview icons.  Some of this is just tweaking the images to
something that I think looks better, but the main gist was
to make the images for hidden windows more distinct.

For many of the common cases, there was little difference
between the visible and hidden cases.  I've removed all of the
color from the hidden images and stripped them down to more
wireframe type images.  Hopefully this makes the visible windows
stand out more clearly from the hidden ones in the treeview.