Correct overdraw issue in the hyperlink controls
The hyperlink controls had an issue where the text would look
odd/darker than it should be.  This was caused by an overdraw
issue where text is being rendered transparently on top of
itself several times.  After the first time, the alpha blending
in the pixels at the edges of the glyphs was blending onto the
earlier results (i.e. it wasn't blending the text onto a clean
background each time).

Unclear why this code was messing around a transparent background
and the hollow brush.  I just deleted those operations and the
overdraw problem goes away, without any other negative effect on
the display of the control.