Details in winspy caption were not always updated
The current window details (HWND and class name) shown in the caption
of the main winspy window were only updated when a window was selected
via the window picker.  The caption was not updated if the active window
was changed through the treeview or the controls on the Windows tab.

Restructured the code so that UpdateMainWindowText is consistently called
any time that spy_hCurWnd changes.