Improve performance, only refresh contents of visible tab
After prior change, winspy no longer hangs outright while trying to
query custom styles from a window in a hung process.  We still will
temporarily hiccup on the SendMessageTimeout calls though, even if
the style tab isn't visible.

Improve this by only refreshing the contents of the visible tab.
There was already code in place that did this for the Class tab,
I just extended that model to all of the tabs.