Handling of NULL/invalid windows for new DPI related fields
Made the recent DPI related fields on the Process and DPI tab
handle NULL and invalid windows in the same fashion as the rest
of the UI.  A NULL window will leave the fields blank, and an
invalid one will report "(invalid window)".

Also changed the main window caption code to report
"(invalid window)" for the window class when the current HWND
isn't valid.