Fix crash when window class name is long
The app would crash if you encountered a window with a long
class name (around 68 characters).

UpdateMainWindowText clamps the window class name that will be
shown in the caption at 70 characters, however at that length
the total size of the formatter string is longer than ach and
_stprintf_s throws an exception.  ach just needs to be a bit
bigger to avoid the crash.