Stop using HWND_DESKTOP for process nodes in tree
Prior to this change, process nodes in the treeview were associated
with the desktop window.  This made for an odd/confusing experience
in that you would select the node for foo.exe, but the Process tab
would show you the details for csrss.exe instead (because the desktop
window is associated with the csrss process).

This changes it so that the process nodes are no longer associated
with any window.  When a process node is selected in the treeview the
Process tab will show the details for that process, otherwise all
the other tabs blank out and show no details (same as the initial
state before any window is ever selected).  Also, right clicking on a
process node in the treeview now shows the process context menu
(terminate etc.) rather than the window one (always on top, etc.).

If you do want to examine the properties of the desktop window,
that is still possible by enabling the (recently added) show desktop
root configuration option.

The main change is that the treeview's item's lParam now points at a
TREENODE struct instead of simply holding an HWND.  This allows
distinguishing between process and window nodes in the tree.  All the
code that directly examined the state of treeview items had to be
tweaked to account for this.  Refactored things such that all the
treeview related code is contained within WinSpyTree.c rather than
being a bit scattered around.

Updated the Process tab to handle the case where the current HWND
is NULL, but there is a process selected.  All of the other details
tabs already showed blank fields in this case.