Command line options to operate in different DPI modes
Added a couple of command line switches to force winspy to run in
different DPI modes.  /pm will cause it to run as per-monitor DPI
aware and /sa will cause it to run as system-aware.  Anything else
will get you a little help message.

The point of this wasn't about making it look better at high DPI,
but rather making it so that you can observe the differences in
window coordinates in the different DPI modes.

By default, winspy runs in DPI unaware mode.  That means that it does
not need to known anything about DPI or attempt to account for it in
the code in any way.  The system automatically stretches the window
content up when running on a high DPI system.  From a usability point
of view this generally fine for a tool like this, the window contents
may be a bit blurry but are always usable.

However, running in unaware mode also means that the system is
reporting window coordinates to winspy in a virtualized 96 DPI
coordinate space.  This gets confusing if you are trying to cross
reference the coordinates/sizes reported by winspy with values
observed in the app itself.  For example, on a 200% DPI monitor,
a DPI aware app could see that a window is at (100,100) but winspy
would report that as (50,50).

These switches only work on Win10 RS1 or later due to the use of the
SetProcessDpiAwarenessContext API to set the DPI mode. If farther
downlevel support is needed, /pm could be supported down to 8.1
via SetProcessDpiAwareness, and /sa could go down to Win7 via
SetProcessDPIAware.

--

When running in high DPI mode, winspy now needs to worry about scaling
its content to account for DPI.  For this app, most things just worked
because the bulk of the UI is built around dialog templates, and the
system does automatically scale dialogs to account for DPI.

I did have to fix a couple of small issues that made usability problems;
the listview colums widths need to be scaled for DPI, and the
WM_MEASUREITEM handler for the owner drawn listboxes needed to be re-worked.
The remaining issue that I'm aware of is that none of the various/icons
bitmaps scale up with DPI.  That is mostly a cosmetic issue.