Drop the help file support.
F1 and/or the Help menu item inserted into the system menu would try
to invoke WinHelp against a winspy.hlp file in the same directory as
the exe.  That help file is long gone and hasn't been in the picture
since the code was put up on GitHub.

Worse, rather than just failing as a silent no-op because the help file
doesn't exist, the call to WinHelp on recent versions of windows kicks
off a browser and takes you to a webpage related to the fact that
support for the old Windows Help format isn't included in windows by
default anymore.

I just deleted the help related code.