Recognize the underlying control types behind winforms class names
In some cases Winforms wraps standard controls with classes of their own.
If we can determine the underlying classname, then use it instead for
the purposes of decoding style names and picking an icon for the treeview.

The underlying class name is just a substring of the winforms one.
For example, "WindowsForms10.SysTabControl32.app.0.fb11c8_r6_ad1" maps
to "SysTabControl32".

Also added the RICHEDIT50W richedit class name variant.