Correct detection of system-aware DPI mode for process tab
Windows 10 version 1803 added a feature which can allow the
'system DPI' associated with different processes to be different.
The method that winspy was using to determine if the DPI context
from the target process was in system-aware mode didn't work if
the other process has a different system DPI than winspy itself.
Changed the logic to use GetAwarenessFromDpiAwarenessContext to
examine the underlying awareness enum and compare it instead.