Horizontal scrolling for caption field stops working after first edit.
You can modify a window's text by editing the value in the caption
fields on the general tab and then hitting the button to the right
of that edit control.  When you do this, the edit control is swapped
out for a combobox.

The initial edit control (IDC_CAPTION1) has the ES_AUTOHSCROLL style,
but the comboxbox that can replace it was missing the corresponding
CBS_AUTOHSCROLL.