Improve auto-update mode by avoiding redundant updates to some controls
Auto-update mode can be annoying because it keeps resetting
the content of the various controls even if nothing changed.
This i=makes a clunky experience if you are trying to interact
with one of those controls at the same time that the refresh
fires and updates it.  For example:

- You are trying to select the text from one of the labels
  (say, with the intent to copy it to the clipboard), the
  refresh will clear the text selection.

- You have a window with more than 8 styles set and you are
  trying to scroll the listbox down to see the bottom of the
  list.  The listbox scroll position will constantly be
  reset to the top.

 This change improves the situation for some of the controls:

 - All the simple labels/edit controls have been changed to skip
   the SetWindowText if the string is exactly the same as what
   is already in the control.  This is done via the new
   SetDlgItemTextEx helper.

 - The Styles tab explicitly tracks what values are being currently
   shown, and doesn't reset the lists unless the styles or the
   selected window have changed.

There are still some controls that this is a problem for, notably
the properties list and the lists on the Windows tab.  Those are
harder to solve, and problems for another day.