Some tweaks to the about dialog
Reworked the language in the about dialog a bit.

I wanted to be able to identify the binaries that I built myself.
Updated the build logic so that if an environment variable called
WINSPY_GITHUB_FORK is defined at build time, then an extra blurb
mentioning that URL is appended at the end of the about dialog
text.

I also bumped the version number to 1.8.  This doesn't really mean
anything, but the last version used on the original catch22.net
site was 1.7, so it makes sense for these post-Catch22 variants
should be numbered at least past that.