Use GetClassLong to query most of the per-class information rather than GetClassInfoEx
Most of the class information can be directly queried via GetClassLong.
As opposed to GetClassInfoEx which requires injecting a thread into
the other process in most cases.

This change switches to using GetClassLong for everything except the
class proc (lpfnWndProc is now the only field in g_WndClassEx
that is used at all).

This makes these fields available in the UI even in the cases where the
thread injection won't work (say, you don't have rights to open a handle
onto the other process, or the other process is stuck in a debugger).

This will make it possible to reduce the scenarios where the thread
injection is attempted, which will be a followup change.

--

Also changed GetRemoteInfo to only attempt WM_GETTEXT when g_fPassword
is true (aka, when the current window is an edit control with the
ES_PASSWORD style).  This call can hang, so it should be avoided when
possible, and the output (g_szPassword) is only used when g_fPassword
is true.