Alt+7 to switch to DPI tab.
There were existing accelerator shortcuts so that Alt+1 would switch to the
first tab (General), Alt+2 would switch to the second tab (Style), etc.
Alt+7 was missing as a means of switching to the DPI tab.

Renamed the accelerator IDs to be a bit more obvious about their purpose,
and moved them to a separate numeric block just because (it didn't matter
right now, but these IDs must be in sequential order as things were there
would not have been room to add a value for an Alt+8 ID if/when one more
tab is added).