Refresh treeview node text/icon when active tab is updated
And refresh the active tab after showing the general poup menu
from eithe the General tab or the treeview.

Combined, these make operations like hiding/showing a window
give some visual feedback about whether or not it actually
happened.  For example, when you hide a window (by unchecking
the 'Visible' menu item) the treeview node for that window will
switch to the grey'ed out version of the icon.  If the General
tab is open, you will see the visible/hidden status update.
If the Styles tab is open, you will see WS_VISIBLE go away.

Added a new image to be used in the treeview for windows that
have have been destroyed.  As you click around in the treeview on
windows that no longer exist, the icon will switch to this
red 'X' which is another indicator that the window is dead.