Add global hotkey to select window under cursor
Added ability to register a global hotkey that once triggered will
'select' the window under the cursor.  This allows you to determine
information about popup menus, tooltips, and things like that that
otherwise are not observable with winspy because the act of
switching focus to winspy would cause them to go away.

Off by default.  You have to turn in on in the options dialog.

I use win+tilde for this.

Note that you can hold down the hotkey and just swipe the mouse
around and it will auto-refresh (by virtue of the hotkey
automatically repeating when held).

If you run multiple instances of winspy, only the first one
tries to register the hotkey.