Fix crashes caused by WM_GETTEXT not null terminating the buffer
WM_GETTEXT does not guarantee that the output buffer is null
terminated.  This behavior differs between control types, but
richedit for sure will leave the buffer unterminated if the text
is longer than the buffer passed to WM_GETTEXT.

This could lead to winspy crashing.  For example, in
CalcNodeTextAndIcon the subsequent call to wcscat_s to add the
closing quote would walk off the end of the buffer and trigger
an explicit failfast.