Improve support for 'extra' styles on the Styles tab.
winspy understands that a handful of controls have an additional class
private style DWORD beyond the normal GWL_STYLE and GWL_STYLEX ones.
On the Styles tab, values from this 3rd style DWORD are listed in the same
listbox as the GWL_STYLEX flags.

This change improves the handling of these extra styles by adding a label
to show the raw hex value and adding an additional '...' button to launch
the style editing dialog for the 'extra' values.

When the class in question has extra styles, the 2nd listbox on the Styles
tab is shortened so that there is room to show the new additional controls.
For windows without the extra styles, the additional controls are hidden and
the listbox is sized/positioned as before this change.

An example of these extra styles can be found in the comctl32 listview
control, where the extra styles are queried via LVM_GETEXTENDEDLISTVIEWSTYLE.
Run regedit and winspy the right hand pane (which is a comctl32 listview)
to see this change in action.

I also made the styles editing dialog (IDD_STYLE_EDIT) slightly wider.
There were a few style names that were wider than the available space in
the listbox control in this dialog.